feat: mostly fixes scanning from the moglabs.
This does not work for the moment. The MOGLabs wavemeter is not able to
trigger on an edges, which makes it trigger multiple acquisitions per
clock cycle of the NI. I tried fixing that using a function generator to
have very short pulses, but triggering is unreliable. i am waiting for
an answer from MOGLabs to sort this out.